% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary2str}
\alias{binary2str}
\title{Binary to string}
\usage{
binary2str(binary_state, sep = ", ", wt = "WT")
}
\arguments{
\item{binary_state}{vector with 0 and 1}

\item{sep}{String. Separator between genes letters}

\item{wt}{String. How to define the wild type}
}
\value{
string with mutated genes
}
\description{
Transform a genotype from binary nomenclature to integer string
}
