% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertraps-process.R
\name{do_HyperTraPS}
\alias{do_HyperTraPS}
\title{Runs HyperTraPS model
This function runs command line tools to call to HyperTraps.
Then processes the output}
\usage{
do_HyperTraPS(
  data = NULL,
  tmp_folder = NULL,
  runs = 1000,
  bi = 50000,
  r = 100,
  seed = 0,
  prob_type = "joint",
  plot = TRUE,
  show_plot = TRUE,
  dry_run = FALSE
)
}
\arguments{
\item{data}{Dataframe with all the cross sectional data}

\item{tmp_folder}{Str. Folder to store and read all the results}

\item{runs}{Int. Number of posterior samples}

\item{bi}{Int. Burn in from MCMC sampler}

\item{r}{Int. HyperTraPS trajectories per transition}

\item{seed}{Int. Ramdom seed for the code to use.}

\item{prob_type}{"joint" or "conditional" probaility type for transitions}

\item{plot}{Boolean. Whether to run the plot command.}

\item{show_plot}{Boolean. Whether to show the plots.}

\item{dry_run}{Boolean. Whether to run the plot command.}
}
\value{
List$edges

List$feature_transitions
}
\description{
HOW SIMULATIONS ARE PERFORMED ()
Greenbury, Sam F. Barahona, Mauricio Johnston, Iain G. 2020
Simulated Walks to Illustrate Order of Acquisition (pages e3-e4)
The inference process above yields inferred posterior distributions on the hypercubic edge weightsW.
We can query these posteriors in a number of ways to gain descriptive and predictive information about the mechanisms generating observed states. 
First, we pro- duce a parsimonious and intuitive representation of the dynamic pathways supported by the inferred posteriors. 
Here, we simulate an ensemble of random walkers generating complete trajectories on hypercubes with sets of transition probabilities sampled from the inferred posterior. 
This ensemble reflects the likely dynamic pathways supported by the dynamic transition model after parameter- ization. 
We simulate an ensemble of random walks in two ways:
 Walk Simulation 1 (WS1), with walkers that run from f0gL to f1gL where a feature is acquired at every time step, and 
Walk Simulation 2 (WS2) which only simulates trajectories corresponding to tran- sitions observed in the dataset. 
In each case, we record every transition between states allowing the construction of a weighted directed graph of all states and transitions encountered. From this graph, the frequency fij with which feature i is gained at step j.
Code for sampling is found in the HypeTraPS repository
HyperTraPS/src/{run_pw.cpp,updater_fast.cpp}
}
