% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn-process.R
\name{inferTheta}
\alias{inferTheta}
\title{Return theta from 
Modified from OncoDB package https://github.com/phillipnicol/OncoBN}
\usage{
inferTheta(df, fit)
}
\value{

}
\description{
Return theta from 
Modified from OncoDB package https://github.com/phillipnicol/OncoBN
}
