% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int2binary}
\alias{int2binary}
\title{Integer to binary}
\usage{
int2binary(int_state, n = NULL)
}
\arguments{
\item{int_state}{Ingeter >=0}

\item{n}{Integer. Number of digits to return}
}
\value{
vector with 0 and 1 with the binary coding
}
\description{
Transform a genotype from integer nomenclature to binary coding
}
