% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-all-methods-minimal.R
\name{plot_DAG_fg}
\alias{plot_DAG_fg}
\title{Plot results from CPMs}
\usage{
plot_DAG_fg(
  x,
  data,
  orientation = "horizontal",
  models = c("OT", "CBN", "DBN", "MCCBN", "MHN", "HESBCN"),
  plot_type = "trans_mat",
  prune_edges = TRUE
)
}
\arguments{
\item{x}{output from the cpm}

\item{data}{Original cross sectional data used to compute the model. Optional.}

\item{orientation}{String. If it not "vertical" will be displayed with an horizontal layout. Optional.}

\item{models}{Output of the CPMs to plot. Current support is for OT, CBN, DBN, MCCBN and MHN Optional.}

\item{plot_type}{String. You can choose between 4 options. Optional.
genotypes: DAG with genotypes transitions
matrix: respresents the transtion matrix as a heatmap
transitions: shows the transitions count between genotypes in HyperTraPS style
             running simulations is needed before this
trans_mat: HyperTraps-like representation of the transition matrix}
}
\description{
By default it create a top row with the DAG of the CPM 
or de transtionRateMatrix for MHN
The bottom row has a custom plot for transition between genotypes
}
\examples{
out <- all_methods_2_trans_mat(dB_c1, do_MCCBN = TRUE)
png("trans_at.png", width = 1000, height = 600, units = "px")
plot_DAG_fg(out, dB_c1, plot_type = "trans_mat")
dev.off()
out2 <- run_all_simulations(out, 100, n_genes = 5)
png("graph.png", width = 1000, height = 600, units = "px")
plot_DAG_fg(out, dB_c1, plot_type = "transitions")
dev.off()
}
