% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-all-methods-minimal.R
\name{plot_genot_fg}
\alias{plot_genot_fg}
\title{Plot transitions between genotypes}
\usage{
plot_genot_fg(
  trans_mat,
  observations = NULL,
  freqs = NULL,
  simplify = TRUE,
  top_edge_per_node = 5
)
}
\arguments{
\item{trans_mat}{transitions matrix to plot: can contain row counts, probabilities... 
Entries will be normalized}

\item{observations}{Original cross sectional data used to compute the model. Optional.}

\item{freqs}{DataFrame with column $Genotype and $Freqs with their frequencies. Optional.}
}
\description{
Plot transitions between genotypes
}
\examples{
png("fluxes.png")
par(mfrow = c(1, 3))
plot_genot_fg(out$MHN_trans_mat, db2, sorted_observations)
title("Transition Matrix", line = -3)
plot_genot_fg(edge_transitions, db2, sorted_observations)
title("Fluxes", line = -3)
dev.off()
}
