% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes_from_trm.R
\name{process_simulations}
\alias{process_simulations}
\title{Process simulations}
\usage{
process_simulations(
  sim,
  n_genes,
  output = c("frequencies", "state_counts", "transitions")
)
}
\arguments{
\item{sim}{list generated with mccbn::sample_genotypes. Relevant
fields are described below}

\item{sim$T_sum_events}{time of events for the mutations of each gene}

\item{sim$obs_events}{data.frame with mutated before the end of the sampling time}
}
\description{
Generate trajectories from simulated data
}
