% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{State.to.Int}
\alias{State.to.Int}
\title{List of sorted genotypes}
\usage{
State.to.Int(x)
}
\arguments{
\item{n_genes}{Ingeter >=0}

\item{sep}{String. Separator between genes letters}

\item{index.return}{Boolean. Wethter to return the indexes}
}
\value{
Vector with the sorted genotypes

Optionally, returns the indexes of the int states
}
\description{
Returns all sorted genotypes for a given number of genes
}
